Add rel=canonical to most old html files.
rel=canonical  is supposed to be helpful for search engine to know
which pages a search engine should promote.

This is _not_ a redirect.

If a file exists and is and is not in a versioned (x.y.z) folder; consider it
as a potential "canonical" target

If a file exist in a versioned doc of matplotlib make its rel=canonical
point to the corresponding candidate.

Example:

    path/to/foo.html -> candidate as being canonical
    3.0.1/path/to/foo.html > set rel=canonical to matplotlib.org/path/to/foo.html

Special case the `examples` and `gallery` folder.
The `example` folder is from old matplotlib version (2.0.2) but still
quite popular on Google; special case.

1) Apply the same rule as above but from withing `examples` toward
`gallery`:

    /gallery/path/to/foo.html -> candidate as being canonical
    /example/path/to/foo.html > set rel=canonical to matplotlib.org/gallery/path/to/foo.html

    This took care of 145 files.

2) Many examples were "moved" during examples -> gallery transition. If
filename is unique in gallery; use this as a potential heuristic for
detecting move:

    /gallery/path/to/verryuniquename.html -> candidate as being canonical as the filename is unique !
    /example/old/location/verryuniquename.html > set rel=canonical to matplotlib.org/gallery/path/to/verryuniquename.html

    This took care of 185 files in example
    Leaving 271 "orphan"

3) Resolve chain of rel-canonical. Fix links that point to example
that themselves point to gallery.

    /gallery/path/to/verryuniquename.html
    /example/old/location/verryuniquename.html ->  matplotlib.org/gallery/path/to/verryuniquename.html
    /2.0.2/example/old/location/verryuniquename.html -> should point to /example/....
                                                       make it point to /gallery/....

No files with existing rel=canonical have been touched. This will
likely only affect pre-3.0 documentation files.