Add an empty Pages deployment
Our repo is too large to deploy to Pages, which is why we use this repo on an external server. However, GitHub keeps trying, and failing, to publish this repo on every commit. For now, I've just kept this around to preserve the redirects from `matplotlib.github.com`.

This change adds a custom workflow that just fills out `CNAME` and publishes that as the "website" on Pages. This avoids both the checkout and the upload of over 10GB that just errors out.